Restructures namespaces and adds event support
Refactors the project to improve code organization and introduces event-driven architecture capabilities.

- Restructures namespaces for better separation of concerns.
- Introduces interfaces for domain and integration events.
- Moves example files to the root directory for clarity.
- Updates package version to 2.0.0 with breaking changes.
