+++ Directory.Build.props
@@ -15,7 +15,7 @@
         <PackageTags>ddd;domain-driven-design;interfaces;extensions;records;classes</PackageTags>
 
         <!-- Version information -->
-        <Version>1.0.0</Version>
+        <Version>2.0.0</Version>
 
         <!-- Documentation settings -->
         <GenerateDocumentationFile>true</GenerateDocumentationFile>
@@ -38,6 +38,7 @@
     <ItemGroup>
         <None Include="$(MSBuildThisFileDirectory)icon.png" Pack="true" PackagePath="\" />
         <None Include="$(MSBuildThisFileDirectory)README.md" Pack="true" PackagePath="\" />
+        <None Include="$(MSBuildThisFileDirectory)CHANGELOG.md" Pack="true" PackagePath="\" />
     </ItemGroup>
     
     <!-- SourceLink setup -->
