+++ Zooper.Lion/Domain/Events/IDomainEvent.cs
@@ -1,4 +1,6 @@
-namespace Zooper.Lion.Interfaces;
+using Zooper.Lion.Common;
+
+namespace Zooper.Lion.Domain.Events;
 
 /// <summary>
 /// Represents a domain event that is used for indicating state changes within a domain (aka. Service).
